+++ pkg/cache/memcached.go
@@ -81,42 +81,3 @@ func (c *MemcachedCache) Fetch(ctx context.Context, keys []string) map[string][]
 	c.hits.Add(float64(len(results)))
 	return results
 }
-
-// mockedMemcachedClient is a mocked memcached client for testing.
-type mockedMemcachedClient struct {
-	cache             map[string][]byte
-	mockedGetMultiErr error
-}
-
-// NewMockedMemcachedClient returns a mocked memcached client.
-func NewMockedMemcachedClient(mockedGetMultiErr error) *mockedMemcachedClient {
-	return &mockedMemcachedClient{
-		cache:             map[string][]byte{},
-		mockedGetMultiErr: mockedGetMultiErr,
-	}
-}
-
-func (c *mockedMemcachedClient) GetMulti(_ context.Context, keys []string) map[string][]byte {
-	if c.mockedGetMultiErr != nil {
-		return nil
-	}
-
-	hits := map[string][]byte{}
-
-	for _, key := range keys {
-		if value, ok := c.cache[key]; ok {
-			hits[key] = value
-		}
-	}
-
-	return hits
-}
-
-func (c *mockedMemcachedClient) SetAsync(_ context.Context, key string, value []byte, _ time.Duration) error {
-	c.cache[key] = value
-	return nil
-}
-
-func (c *mockedMemcachedClient) Stop() {
-	// Nothing to do.
-}
