+++ pkg/cache/memcached_test.go
@@ -70,7 +70,7 @@ func TestMemcachedIndexCache(t *testing.T) {
 
 	for testName, testData := range tests {
 		t.Run(testName, func(t *testing.T) {
-			memcached := NewMockedMemcachedClient(testData.mockedErr)
+			memcached := newMockedMemcachedClient(testData.mockedErr)
 			c := NewMemcachedCache("test", log.NewNopLogger(), memcached, nil)
 
 			// Store the postings expected before running the test.
@@ -86,4 +86,43 @@ func TestMemcachedIndexCache(t *testing.T) {
 			testutil.Equals(t, float64(len(testData.expectedHits)), prom_testutil.ToFloat64(c.hits))
 		})
 	}
+}
+
+// mockedMemcachedClient is a mocked memcached client for testing.
+type mockedMemcachedClient struct {
+	cache       map[string][]byte
+	getMultiErr error
+}
+
+// newMockedMemcachedClient returns a mocked memcached client.
+func newMockedMemcachedClient(getMultiErr error) *mockedMemcachedClient {
+	return &mockedMemcachedClient{
+		cache:       map[string][]byte{},
+		getMultiErr: getMultiErr,
+	}
+}
+
+func (c *mockedMemcachedClient) GetMulti(_ context.Context, keys []string) map[string][]byte {
+	if c.getMultiErr != nil {
+		return nil
+	}
+
+	hits := map[string][]byte{}
+
+	for _, key := range keys {
+		if value, ok := c.cache[key]; ok {
+			hits[key] = value
+		}
+	}
+
+	return hits
+}
+
+func (c *mockedMemcachedClient) SetAsync(_ context.Context, key string, value []byte, _ time.Duration) error {
+	c.cache[key] = value
+	return nil
+}
+
+func (c *mockedMemcachedClient) Stop() {
+	// Nothing to do.
 }
